# tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "apex-admin",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/files": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "上传文件",
        "description": "上传文件 — multipart/form-data（SPEC 19.2 / 19.3）.\n\nSPEC 19.2:\n  - 限制单文件大小。\n  - 限制单次上传数量。\n  - 分块流式写入，禁止一次性读取整个文件到内存。\n  - 使用白名单校验允许的文件类型。\n\n每个文件独立处理，部分失败时已成功的文件仍然返回。",
        "operationId": "upload_files",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileMetadataResponse"
                  },
                  "title": "Response Upload Files"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "file"
        ],
        "summary": "查询当前用户文件列表",
        "description": "查询当前用户的文件列表 — SPEC 19.4.",
        "operationId": "list_my_files",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileMetadataResponse"
                  },
                  "title": "Response List My Files"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{fileId}": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "查询文件详情",
        "description": "查询文件详情 — SPEC 19.1.",
        "operationId": "get_file",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文件 ID",
              "title": "Fileid"
            },
            "description": "文件 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "删除文件",
        "description": "删除文件 — SPEC 19.3 / 19.4.\n\nSPEC 19.3: \"删除前必须在事务中确认没有活动业务引用并更新为 DELETING\"。\nSPEC 19.3: 物理删除延迟至少 7 天（reconcile 在 TASK-026 执行）。\n\nSPEC 19.4: 上传者或拥有文件管理权限的管理员可删除。",
        "operationId": "delete_file",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文件 ID",
              "title": "Fileid"
            },
            "description": "文件 ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{fileId}/download": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "下载文件",
        "description": "下载文件 — SPEC 19.2 / 19.3 / 19.4.\n\nSPEC 19.3: 仅 READY 文件可下载。\nSPEC 19.4: 上传者或拥有文件管理权限的管理员可下载。\nSPEC 19.2: 响应头安全处理原始文件名（RFC 5987）。\nSPEC 19.2: 下载接口防止未授权枚举文件（不区分 404 和 403）。",
        "operationId": "download_file",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文件 ID",
              "title": "Fileid"
            },
            "description": "文件 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_files": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "上传的文件列表"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_files"
      },
...
      "FileMetadataResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "originalName": {
            "type": "string",
            "title": "Originalname"
          },
          "storageName": {
            "type": "string",
            "title": "Storagename"
          },
          "sizeBytes": {
            "type": "integer",
            "title": "Sizebytes"
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype"
          },
          "fileExtension": {
            "type": "string",
            "title": "Fileextension"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "uploadedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadedby"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "deletingEnteredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletingenteredat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "originalName",
          "storageName",
          "sizeBytes",
          "contentType",
          "fileExtension",
          "sha256",
          "status",
          "uploadedBy",
          "createdAt",
          "updatedAt"
        ],
        "title": "FileMetadataResponse",
        "description": "文件元数据响应模型 — SPEC 9.3 / 19.1."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque",
        "description": "粘贴登录接口返回的 access_token（无需手写 Bearer 前缀）。认证依赖从 Authorization: Bearer <token> 提取不透明 Access Token（SPEC 12.1）。"
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    }
  ]
}